requirements.txt
alabaster==0.7.2
autoprotocol==2.0.2
Babel==1.3
click==3.3
numpy==1.9.2
Pint==0.6
py==1.4.26
pycallgraph==1.0.1
Pygments==2.0.2
pymongo==2.8
pytz==2014.10
PyYAML==3.11
requests==2.6.0
scipy==0.15.1
simplejson==3.6.5
six==1.9.0
snowballstemmer==1.2.0
statistics==1.0.3.5
tox==1.9.0
transcriptic==1.0.1
virtualenv==12.0.7